requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
python-decouple>=3.8,<4.0
whitenoise>=6.5,<7.0
gunicorn>=21.2,<23.0
psycopg2-binary>=2.9,<3.0
stripe>=7.0,<8.0
